+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project uses [Semantic Versioning 2.0.0](http://semver.org/).
 
 ## main
 
+- ADDS: Domains Service
 - ADDS: OAuth Service and Error handling
 - ADDS: Accounts Service
 - ADDS: Identity Service